Stop escaping = in completion. fixes #18
Previously, bash-completion.el used shell-quote-argument from subr.el to
quote unsafe characters, which has a rather restrictive view of which
characters can be considered safe. This commit replaces
shell-quote-argument with custom code, and introduces
bash-completion-special-chars. This is, for now, the same regexp as in
shell-quote-argument, with = whitelisted.

Moving away from shell-quote-argument also avoids possible issues with
quoting on windows: since we know we're communicating with bash, there's
no reason to quote characters in a way that's platform-specific.