Add bash-completion-dynamic-complete-nocomint for issue #13
This commit introduces a function that can safely be called outside of a
comint buffer. It is passed the region to complete and returns the same
result as bash-completion-dynamic-complete without accessing any
comint-specific function.

The refactoring removes bash-completion-dynamic-complete-0 and
bash-completion-dynamic-try-wordbreak-complete. Obsolete wrappers are
left for now, to avoid making a breaking change, for now.