Add / after directory names - even if they contain spaces or special characters, escaped with \\.
When completing filenames with spaces, emacs bash completion escapes
spaces and other special characters with backslashes. So when
completing:
  ls ~/Documents/My
to the directory "My Books", emacs bash completion completes it to:
  ls ~/Documents/My\\ Book/
with this fix, the / is there. Without it, emacs bash completion would
look for a directory called "My\\ Book" - fail to find it and not add
the backslash.