Preserve original, quoted version of the string being completed when returning results.
One of the first things that bash-completion-dynamic-complete-nocomint
does is parse and tokenize the current line, with
bash-completion-process-tokens. compgen then gets the parsed result and
returns matches for that parsed result.

In some cases, running compgen on the parsed result resulted in
completed words not starting with the word to be completed, as found in
the buffer. Such results were ignored by the shell completion.

Example:
 ls ~/Documents/Modes\ d\'emplois/M

compgen would see "~/Documents/Modes d'emplois" without backslashes and
return, for example "~/Documents/Modes d'emplois/MXX.pdf" - but this
doesn't match ~/Documents/Modes\ d\'emplois/

With this fix, the original, unparsed version of the string is passed
down to bash-completion-fix, so that it can postprocess the completion
candidates returned by compgen and guarantee that they match exactly
with the buffer content.