Detect directories when expanding commands and add a slash instead of a space. fixes #11
Before this commit, the parameter -S ' ' was passed to compgen when
expanding a command, which means that it would always append a space.
Adding a space is useful when expanding a command, but annoying when
expanding a directory.

With this commit, the space is added by bash-completion.el instead of
bash commands, allows checking for matching directory.