Fixes strange behavior when completing --.
Completing -- in commands where no completion has been defined used to
default to file completion, so for example:
ls --<complete>
would complete to
ls --bar
if there was a file bar in the current directory.

This is because the string to be completed was passed to compgen,
and when compgen sees --, it just disables parsing of further
parameters. It didn't happen when using custom completion, because
bash-completion.el already passed -- in this case.

This change adds the missing -- for default and command completion, to
avoid issue when completing -- or files starting with -.

Mentioned in issue #13