Fix occasional escaping issue with candidates containing quotes.
bash-completion-escape-candidate used to attempt detect quoted
strings. The idea was that programmable completion could return quoted
strings. It's impossible, though, functions fill in results as bash
strings in COMPRELPY; they don't print them.

The commit f3b11e1 broke that idea by calling
bash-completion-escape-candidate only on the substring that follows
the stub to be completed. Since that change, single or double quotes
anywhere in the strings could end up not being escaped when they
should have been, just because they happened to be the first part of
the string to append to the completion candidate.

Example:
 cat Documents/Modes\ d<point>
Now completes to:
 cat Documents/Modes\ d\'emplois<point>
But would, before this commit, complete to:
  cat Documents/Modes\ d'emplois<point>