Make wordbreak completion work with bash-completion-enable-caching.
With this commit, wordbreak completion works even when caching is
enabled, though it will look slightly different, as completion appears
to start at the beginning of the argument, instead of the beginning of
the word.

This commit also extends integration test to cover the case where
bash-completion-enable-caching is set.

bash-completion-enable-caching is kept as an option, since it changes
visible behavior.

fixes #30