Rely on compgen for default completion, instead of doing it with another call to compgen. "-o default" implies filenames option.
Also removes the now useless customization option
for forcing or disabling default and the unnecessary one for
filenames. Only forcing or disabling nospace remains.

This change should result in a behavior closer to bash, while making
fewer calls to compgen, so speeding up completion in general. This
should be especially visible when using tramp

Issue #19