Add a new option to disable the use of separate processes to perform completion
The drawback of using separate processes to do the completion is that
the completion process is not aware of any changes made to bash in the
current buffer. Setting the new option
`bash-completion-use-separate-processes' to nil, remove this
limitation by using the bash associated with the current buffer to
perform the completion.