Make /etc/bash_completion scripts work with escaped spaces.
/etc/bash_completion uses dynamic completion for everything, instead
of default completion. Completing files with escaped spaces didn't
work with these scripts, because they rely on compgen -f, whose
behavior is changes depending on whether readline is enabled - in a
normal terminal - or not - in Emacs.

This change works around the problem by feeding pre-unquoted strings
to the scripts.