Attempt to make in-process execution more robust.
When use-separate-processes is nil, this change waits for a special
string, added by bash-completion-send, instead of the prompt regexp.
This should be more robust, as prompts can change and the prompt
regexp might easily match output.

This change also removes uses of comint-last-prompt. This was an
attempt at making the prompt search more reliable, but won't work if
the prompt changes, such as a prompt that includes the history line
count.