Respect completion-ignore-case setting from .inputrc.
This change reads the setting completion-ignore-case set in .inputrc
and use it to set the Emacs variable completion-ignore-case to make
completion accept matches whose case don't match.

Case-insensitive completion only works on bash 4, not bash 3, as this
approach relies on compgen following the setting in .inputrc, which
only happens starting with bash 4.

This change, together with the previous changes accepting completion
output that don't start with the original word, fix issue #7.