Don't let Emacs post-filter completions built by bash.
With this change, the completions built by bash are offered to the
user even if they don't match the string to be completed.

Before this change, the dynamic completion table would call
try-completion/all-completion/test-completion, which would then filter
the result according to the current completion style, which makes no
sense as bash produces complete completions.

Not allowing post-processing of the completion also means that there's
no need to worry about the value of completion-ignore-case anymore on
the Emacs side; bash-completion.el just serves whatever bash returns.

Note that post-filtering is still possible when using
bash-completion-dynamic-complete-nocomint with the argument
dynamic-table left nil. In this scenario, the completion engine is
free to process the results as it wants.