completion-table-with-cache captures and restore buffer and variables.
With this change, completion-table-with-cache restores the original
buffer. This is useful as here's no guarantee of what the current
buffer is when the dynamic table function is called, so enforce it
before calling bash-completion--complete, to be sure.

This change also restores the current directory as well as the value
of bash-completion-use-separate-processes and bash-completion-nospace,
as both these are often overridden temporarily in a let section.