Avoid interference from PROMPT_COMMAND in single process mode.
This change avoids letting the commands from PROMPT_COMMAND and PS1
run during a completion, when doing completion in process. This is
necessary as PROMPT_COMMAND can do absolutely anything. Also, it's not
expected to run except as a result of a user action.

This change helps avoid relying on the prompt detection regexp
anywhere in the code; the only prompt that's detected is the fake one
setup by bash-completion.el.

Also switches the special functions and variables set up by
bash-completion.el from using the prefix __bash_completion, which is
confusing when you're in bash anyways, to using the prefix
__emacs_completion.