Remove the fallback to a separate process.
Falling back to using a separate process when there's a failure in the
single-process solution hasn't worked out so well in practice. The
problem is that when there's an error, it's usually after a noticeable
timeout, so the resulting behavior is confusing even in the cases
where the fallback actually works. Better to fail clearly and
hopefully get errors reported - and fixed.