Let status code of compgen through when in single-process mode.
(bash-completion-send) is supposed to return the status of the
command. However, when running in single-process mode, a command was
executed after the main command, so the status returned was the status
of that second command.

This change transforms __emacs_complete_post_command into
__bash_complete_pre_command.

This change also silences errors from the history command, which is
how this issue came to light, as errors from history made it look like
the compgen command failed.