Avoid escaping all non-ASCII characters.
This change updates the set of special chars to not include non-ASCII
chars.

The previous set of special chars to escape was previously defined by a
list of characters *not* to escape, which included only ASCII
characters. This meant that anything non-ascii would be escaped, and
unnecessarily so since these characters are not special for bash.