Test and fix the dynamic table function that's returned.
This change attempts to make the dynamic table function behave in a way
that more closely matches:
  https://www.gnu.org/software/emacs/manual/html_node/elisp/Programmed-Completion.html

It doesn't follow the spec in all ways, as it assumes that all-complete
and try-complete understand how the completion function works, whereas
completion is arbitrary with bash.

To work around that problem, the dynamic table deviates from spec when
passed as a string the original, unmodified word that was completed and
returns returns all completion, without filtering them. Otherwise, it
follows spec.

Before this change, the completion function never really followed spec,
which was ok for the default completion functions, but confused custom
completion functions.

This is a follow-up to issue #43 and should allow helm completion to
behave more sanely when backed by bash-completion.