Detect non-responsive processes more quickly.
With this change, bash-completion-send first sends a short echo command,
just to see whether the process answers properly, so that, in the case
where the process is not able to respond or not a bash process,
bash-completion fails quickly.

This makes issue #44 less annoying but doesn't actually solve the
problem yet.