Remove the need for a special setup in single-process mode.
Before this change bash-completion-send would rely on custom functions,
defined during setup, when bash-completion-use-separate-processes is
nil. This meant that if, for some reason setup hadn't run, communication
with the Bash process would just not be possible and the completion
would freeze.

This typically happened when a bash subprocess is created, as Emacs
would think that setup was run when it wasn't. With this change,
completion still doesn't work when called from bash subprocessed, but it
at least doesn't freeze.

This change required changing the prompt (PS1) used when communicating
with bash to only include printable characters, to avoid issues when -o
vi or -o emacs is set in the configuration.