Run bash-completion--setup-bash-common when needed by bash.
Before this change, bash-completion.el would guarantee that
bash-completion--setup-bash-common is run exactly once by flagging the
process.

This would fail in the case where a bash subprocess was created, as the
functions and options set during setup in the main Bash process would
not be available in the subprocess.

This change runs bash-completion--setup-bash-common whenever Bash
reports not __emacs_complete_wrapper, so it's safe to run completion
from Bash subprocesses.

This change, together with the previous commit, should address
isse #58, as there's no need to worry about calling bash-completion-refresh
at the right time anymore.