Fail fast when completion is run against a process that is not bash.
Before this change when the currently active command-line tool didn't
answer as expected, bash-completion.el would wait for
bash-completion-command-timeout for a response.
bash-completion-command-timeout needs to be large, as some completion
take a long time to execute.

With this change, Bash sends a special string back before executing
completion. Waiting for that string can be short, as it's cheap and fast
for Bash to execute.

Worst case, if the command-line tool isn't Bash anymore, the user now
waits up to 600ms for an answer from Bash before getting back an error
message. That new short timeout can be configured by customizing the
value of bash-completion-short-command-timeout.