Move appending of / to directories from Emacs to Bash.
Before this change, completion results were post-processed on the Emacs
side and a / was appended to anything that looked like a directory.

This didn't work in the case where Emacs know the host, user and
environment the bash process runs in, which is easy to get to by running
ssh or sudo from inside the shell.

With this change, post-processing happens on the Bash side, so the're no
risk of mismatch and no need for bash-completion.el to track the current
directory anymore.

Implementation turned out to be tricky. You'd normally do
post-processing of compgen output using a pipe, but the the functions
called by compgen sometimes modify the environment. This would have been
broken since pipes create subshells. This change uses compgen ... <
<(function), to post process, which leaves compgen in the current shell.

This change also:

- declares some temporary shell variables used in functions "local".
- avoids cases where complete -p was run more than once for a single
  expansion in bash-completion--customize