Have bash-completion-debug report all of BASH_VERSION.
Before this change, when not in separate process mode, only the major
and minor version of the Bash process was reported, in an attempt at
avoiding issues with BASH_VERSION containing unexpected characters.

This change uses BASH_VERSION, instead, as the content of BASH_VERSION
is reasonably well defined and constrained.