Avoid executing anything in case __ebcpre is not defined.
In single-process mode, to detect communication with an un-configured
Bash process, bash-completion.el always starts with a check for its
internal function __ebcpre (type __ebcpre) and configures the Bash
process if that doesn't succeed.

Before this change, the case where type __ebcpre failed was correctly
detected and reported by Bash but the command was executed anyways - and
the result usually thrown away.

This was wasteful, but should have been harmless. However, in some
cases, at least when expanding a command name first thing in a buffer,
that command happened to cause Bash to exit or crash.

With this change, the intended command isn't executed anymore when type
__ebcpre fails, which avoids crashing Bash.

fixes #63