Avoid executing PROMPT_COMMAND when doing completion.
Before this change, PROMPT_COMMAND was sometimes allowed to execute
during the first completion in a Bash process. This caused confusion in
the case where PROMPT_COMMAND does more than just printing data.

With this change, PROMPT_COMMAND is guaranteed to be stored and then
cleared in the very first command set to a Bash process, together with
PS1.

Note that this lengthens the commands sent to Bash, to the point where
they were getting too long when defining functions, so initial function
definition has been split into two Bash commands instead of one.

For background, see discussion on issue #63