Fix strange issue with scripts using < <(cmd).
Before this change, bash-completion.el would check the version of Bash
and if the version was recent enough, use
 compgen ... > >(__ecbfixdirs); wait $!
to post-process the output of compgen.

This seems to be causing strange behaviors in scripts using
  ... < <(cmd)

This change always falls back to the unoptimized __ecbfixdirs, which
uses pipes and doesn't seem to trigger this issue. There's most likely a
better solution, but until then, this seems to work.

Issue #74