Fix Contributing section in README.md
Before this change, the contributing section would define link with the syntax [text][id] and the id at the end of the file as reference. This doesn't seem to work (anymore?)

This change updates the section to use normal links [text](url) and mention opening a new discussion instead of only creating issues.