fix: Recover $? after completion.
Before this change, the value of $? was lost when doing a completion as
it required running a command, so $? became the status code of the
completion command.

So if you typed:

> false
> ech<TAB> $?

You would get 0 instead of 1, set by false.

This change stores the value of $? first thing before executing any
command, then have __ebcret restore it. The status code that bash
completion, the one that's embedded in the next prompt, remains the
status code of the completion command, but $? is the status code of the
last user command, before completion was run.

issue #77