fix: Recover PS1 before command execution.
Before this change, when doing completion in the same process PS1 was
modified to be able to detect the end of the completion command output
and it was recovered later on from PROMPT_COMMAND, just *after*
executing the user command, before building the next prompt.

The effect was not visible, unless the user command included $PS1, so if
you did:

> echo $PS1

and ran completion while editing that command, the PS1 that would be
output would be the one from bash-completion.

This change switches to another approach for recovering the prompt that
allows recovering it before executing the user command: a DEBUG trap is
registered which restores PS1 and PROMPT_COMMAND if the previous command
included __ebrcet. This change also adds __ebcret after the last command
issued by completion.

This way, the last command issued by completion uses the fake prompt,
but the command run just after that sees the real prompt in its
variable.

issue #77