fix: Update prompt after a syntax error.
Before this change, the prompt may be left to the special bash prompt
after a syntax error, because Bash won't execute the DEBUG trap at all
in such a scenario. At least recent ones won't; that might depend on the
exact version of Bash.

This change switches to a new scheme that recovers PS1 after bash has
executed commands without relying on a trap. This means that PS1 may be
printed, but after the special bash-completion marker, so the string
will be skipped.

PROMPT_COMMAND is still recovered as it was before this change, using a
DEBUG trap, to avoid slow prompt command making completion unusable. The
drawback of this approach is that that PS1 may not updated after a bash
syntax error as it would normally.

issue #79