fix: Avoid commands in PS1 causing issues.
With the fix of issue #79, PS1 is displayed after the bash completion
mark, but later on ignored. This doesn't work if PS1 executes commands,
as the command execution will trigger the TRAP.

This change evaluates PS1 beforehand, so it won't confuse the trap. With
this approach, an outdated PS1 may be displayed after a Bash syntax
error, but this is temporary, as executing any command fixes it.

issue #80